more work towards multiview
seems like the issue is that each render item stores model view and matrix in a single buffer, so overwriting it for teh second view will always take precedence as rendering is not synchronous. if we split out the camera members (projection and view matrix) to a new struct, can we change just that buffer between draw calls?